// types.ts
export interface Question {
  id: number;
  text: string;
  options: string[];
  correctAnswerIndex: number;
  prize: string;
  difficulty: 'easy' | 'medium' | 'hard';
  subject: string;
  chapter: string;
  lesson: string;
  isAi?: boolean;
}

export interface MoneyTier {
  level: number;
  amount: string;
  value: number;
  isSafeHaven: boolean;
}

export interface Lifelines {
  fiftyFifty: boolean;
  askAudience: boolean;
  callFriend: boolean;
}